move tui diff storage to tui crate
