+++ crates/warp_tui/Cargo.toml
@@ -33,6 +33,7 @@ name = "warp-tui-stable"
 path = "src/bin/stable.rs"
 
 [dependencies]
+ai.workspace = true
 anyhow.workspace = true
 async-fs.workspace = true
 channel_versions.workspace = true
@@ -49,6 +50,7 @@ parking_lot.workspace = true
 pathfinder_color.workspace = true
 pathfinder_geometry.workspace = true
 serde_json.workspace = true
+similar.workspace = true
 string-offset.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
@@ -58,11 +60,14 @@ warp = { workspace = true, features = ["tui"] }
 warp_channel_config.workspace = true
 warp_core.workspace = true
 warp_editor.workspace = true
+warp_files.workspace = true
 warp_terminal.workspace = true
+warp_util.workspace = true
 warpui.workspace = true
 warpui_core = { workspace = true, features = ["tui"] }
 
 [dev-dependencies]
+tempfile.workspace = true
 # `test-util` exposes `Appearance::mock()`, used to register the `Appearance`
 # singleton that the editor-backed TUI input view depends on in tests.
 warp_core = { workspace = true, features = ["test-util"] }
