+++ crates/warp_tui/src/lib.rs
@@ -27,6 +27,7 @@ mod terminal_session_view;
 mod test_fixtures;
 mod transcript_view;
 mod tui_block_list_viewport_source;
+mod tui_diff_storage;
 mod tui_file_edits_view;
 
 pub use root_view::RootTuiView;
